関数の統合
