--- ingredients/Meridian_TWIN/Meridian_TWIN_ESP32/src/mrd_util.h
@@ -1,114 +1,0 @@
-#ifndef __MERIDIAN_UTILITY_H__
-#define __MERIDIAN_UTILITY_H__
-
-// ヘッダファイルの読み込み
-#include "config.h"
-#include "main.h"
-
-//================================================================================================================
-// Utility ごく小規模な汎用関数
-//================================================================================================================
-
-/// @brief 配列の中で0以外が入っている最大のIndexを求める.
-/// @param a_arr 配列
-/// @param a_size 配列の長さ
-/// @return 0以外が入っている最大のIndex. すべて0の場合は1を反す.
-int mrd_max_used_index(const int a_arr[], int a_size) {
-  int max_index_tmp = 0;
-  for (int i = 0; i < a_size; ++i) {
-    if (a_arr[i] != 0) {
-      max_index_tmp = i;
-    }
-  }
-  return max_index_tmp + 1;
-}
-
-/// @brief 指定された位置のビットをセットする(16ビット変数用).
-/// @param a_byte ビットをセットする16ビットの変数.参照渡し.
-/// @param a_bit_pos セットするビットの位置(0から15).
-/// @return なし.
-inline void mrd_setBit16(uint16_t &a_byte, uint16_t a_bit_pos) { a_byte |= (1 << a_bit_pos); }
-
-/// @brief 指定された位置のビットをクリアする(16ビット変数用).
-/// @param a_byte ビットをクリアする16ビットの変数.参照渡し.
-/// @param a_bit_pos クリアするビットの位置(0から15).
-/// @return なし.
-inline void mrd_clearBit16(uint16_t &a_byte, uint16_t a_bit_pos) { a_byte &= ~(1 << a_bit_pos); }
-
-/// @brief 指定された位置のビットをセットする(8ビット変数用).
-/// @param value ビットをセットする8ビットの変数.参照渡し.
-/// @param a_bit_pos セットするビットの位置(0から7).
-/// @return なし.
-inline void mrd_setBit8(uint8_t &value, uint8_t a_bit_pos) { value |= (1 << a_bit_pos); }
-
-/// @brief 指定された位置のビットをクリアする(8ビット変数用).
-/// @param value ビットをクリアする8ビットの変数.参照渡し.
-/// @param a_bit_pos クリアするビットの位置(0から7).
-/// @return なし.
-inline void mrd_clearBit8(uint8_t &value, uint8_t a_bit_pos) { value &= ~(1 << a_bit_pos); }
-
-/// @brief 列挙型(L,R,C)から文字列を取得する関数.
-/// @param a_line 列挙型 enum UartLine
-/// @return 列挙型の内容に応じて文字列"L","R","C"返す.
-const char *mrd_get_line_name(UartLine a_line) {
-  switch (a_line) {
-  case L:
-    return "L";
-  case R:
-    return "R";
-  case C:
-    return "C";
-  default:
-    return "Unknown";
-  }
-}
-
-//------------------------------------------------------------------------------------
-//  タイムアウト監視用タイマー
-//------------------------------------------------------------------------------------
-// タイマーの開始状態を保持するための静的変数
-static unsigned long timeout_start = 0;
-static bool flg_timer_started = false; // タイマーが開始されたかどうかのフラグ
-
-/// @brief 指定されたミリ秒のタイムアウトを監視する. mrd_timeout_resetとセットで使う.
-/// @param a_limit タイムアウトまでの時間(ms)
-/// @return タイムアウトでtrueを返す.
-bool mrd_timeout_check(unsigned long a_limit) {
-  // タイマーが開始されていない場合は現在の時間を記録してタイマーを開始
-  if (!flg_timer_started) {
-    timeout_start = millis();
-    flg_timer_started = true; // タイムアウト監視開始フラグをアゲる
-  }
-
-  unsigned long current_time = millis(); // 現在の時間を取得
-
-  if (current_time - timeout_start >= a_limit) { // 指定された時間が経過しているかチェック
-    flg_timer_started = false;                   // タイムアウト監視開始フラグをサゲる
-    return true;                                 // 指定された時間が経過していれば true を返す
-  }
-
-  return false; // まだ時間が経過していなければ false を返す
-}
-
-/// @brief タイムアウト監視開始フラグをリセットする. mrd_timeout_checkとセットで使う.
-void mrd_timeout_reset() {
-  flg_timer_started = false; // タイマーをリセットして次回の呼び出しに備える
-}
-
-//------------------------------------------------------------------------------------
-//  meriput / meridimへのデータ書き込み
-//------------------------------------------------------------------------------------
-
-/// @brief meridim配列のチェックサムを算出して[len-1]に書き込む.
-/// @param a_meridim Meridim配列の共用体. 参照渡し.
-/// @return 常にtrueを返す.
-bool mrd_meriput90_cksm(Meridim90Union &a_meridim, int len = 90) {
-  int a_cksm = 0;
-  for (int i = 0; i < len - 1; i++) {
-    a_cksm += int(a_meridim.sval[i]);
-  }
-  a_meridim.sval[len - 1] = short(~a_cksm);
-  return true;
-}
-
-#endif //__MERIDIAN_UTILITY_H__
